removing complexity from the code
